Move safe TSV reader to library
